pokecertify/requirements.txt
fastapi
uvicorn[standard]
python-multipart
requests
qrcode
reportlab
pillow
gradio
web3
modal
